+++ src/plone/app/robotframework/i18n.py
@@ -1,12 +1,13 @@
 # -*- coding: utf-8 -*-
-import os
-
-from Products.CMFCore.utils import getToolByName
 from plone.app.robotframework.remote import RemoteLibrary
 from plone.app.robotframework.utils import disableCSRFProtection
+from plone.registry.interfaces import IRegistry
+from zope.component import getUtility
 from zope.component.hooks import getSite
 from zope.globalrequest import getRequest
 from zope.i18n import translate
+
+import os
 
 
 class I18N(RemoteLibrary):
@@ -14,12 +15,20 @@ class I18N(RemoteLibrary):
     def set_default_language(self, language=None):
         """Change portal default language"""
         disableCSRFProtection()
-        portal = getSite()
-        portal_languages = getToolByName(portal, 'portal_languages')
         if language is None:
             language = os.environ.get('LANGUAGE') or 'en'
-        setattr(portal, 'language', language)
-        portal_languages.setDefaultLanguage(language)
+        try:
+            from Products.CMFPlone.interfaces import ILanguageSchema
+        except ImportError:
+            portal = getSite()
+            return portal.portal_properties.site_properties.setProperty(
+                'default_language',
+                language,
+            )
+        else:
+            registry = getUtility(IRegistry)
+            settings = registry.forInterface(ILanguageSchema, prefix='plone')
+            settings.default_language = language
 
     def translate(self, msgid, *args, **kwargs):
         """Return localized string for given msgid"""
@@ -31,13 +40,15 @@ class I18N(RemoteLibrary):
 
         mapping = {}
         for key, value in kwargs.items():
-            if not key in ('target_language', 'domain', 'default'):
+            if key not in ('target_language', 'domain', 'default'):
                 mapping[key] = value
         if kwargs.get('target_language'):
             return translate(
-                msgid, target_langauge=kwargs.get('target_language'),
+                msgid,
+                target_langauge=kwargs.get('target_language'),
                 domain=kwargs.get('domain') or 'plone',
-                default=kwargs.get('default') or msgid, mapping=mapping)
+                default=kwargs.get('default') or msgid, mapping=mapping
+            )
         else:
             # XXX: Should self.REQUEST be replaced with
             # zope.globalrequest.getRequest()?
@@ -45,4 +56,5 @@ class I18N(RemoteLibrary):
             return translate(
                 msgid, context=request,
                 domain=kwargs.get('domain') or 'plone',
-                default=kwargs.get('default') or msgid, mapping=mapping)
+                default=kwargs.get('default') or msgid, mapping=mapping
+            )